wip: rename vydia -> sitemate
