+++ android/src/main/java/com/sitemate/uploader/GlobalRequestObserverDelegate.kt
@@ -1,4 +1,4 @@
-package com.vydia.RNUploader
+package com.sitemate.uploader
 
 import android.content.Context
 import android.util.Log
