+++ android/src/main/java/com/sitemate/work/UploadWorker.kt
@@ -1,12 +1,12 @@
-package com.vydia.RNUploader
+package com.sitemate.uploader
 
 import android.content.Context
 import androidx.work.Worker
 import androidx.work.WorkerParameters
-import com.vydia.RNUploader.extensions.PARAM_KEY_NOTIF_CONFIG
-import com.vydia.RNUploader.extensions.PARAM_KEY_TASK_PARAMS
-import com.vydia.RNUploader.extensions.toUploadNotificationConfig
-import com.vydia.RNUploader.extensions.toUploadTaskParameters
+import com.sitemate.uploader.extensions.PARAM_KEY_NOTIF_CONFIG
+import com.sitemate.uploader.extensions.PARAM_KEY_TASK_PARAMS
+import com.sitemate.uploader.extensions.toUploadNotificationConfig
+import com.sitemate.uploader.extensions.toUploadTaskParameters
 
 class UploadWorker(val context: Context, params: WorkerParameters): Worker(context, params) {
