+++ example/RNBackgroundExample/package.json
@@ -20,7 +20,7 @@
   "dependencies": {
     "react": "18.2.0",
     "react-native": "0.71.13",
-    "react-native-background-upload": "github:site-mate/react-native-background-upload#cda58cf8ab1878bfc78afc800678f101a07f234a",
+    "react-native-background-upload": "github:site-mate/react-native-background-upload#eb4fa54d1cee009d19144c256e5c9768979dd4e5",
     "react-native-fs": "^2.20.0",
     "react-native-image-crop-picker": "^0.40.3"
   },
